+++ CHANGELOG.md
@@ -4,6 +4,8 @@ hook_update_deploy_tools 7.x-1.x **-**-****
 
 hook_update_deploy_tools 7.x-1.24 Mar 17, 2017
 -----------------------------------------------
+* Issue #2861596: Create method to delete a Vocabulary Term
+  https://www.drupal.org/node/2861596
 * Issue #2603886: Import and export taxonomy term from a txt file'
   https://www.drupal.org/node/2603886
 * Add Vocabulary load methods.